modify model files
